Improved Comments, Split tasks
During this Improvement, I decided to rewrite Comments, for a better understanding and get rid of unnecessary comments.

Understanding of Running each task from Main  " Source.cpp" File and writing Code between Header and its source code.

In the header file, we make calls for classes  and Write our variables.
And in the source is where the code is executed.

Further possible improvements, is to Finish both Exercise 1 and 3, and further improve the Code implemented and Move forward to getting through with the rest of the Exercises .