Task 3 Updates, Comment Rewording.
Completed Task 3, where It takes
Input from the User " Keyboard" With a getline, and assigns to a string, Then goes into a loop which splits up the Elements in that string  Into separate words And Stores Each word into a Vector Array " List "

Then Checks If the List is == 0  Empty then calls Return.

Prints last word in list and Prints them In reversed Order.