Task4, Small changes to task3.
Completed Task 4.
First I let the user Make an Array list size, then ask them to put in the elements "Integers "

t4.getuserinput()  - Array Size Makeing and Input to the List.

Tell user  Which number they would be looking for from that made List,

Input the Key to , key  ( placeholder element to store their input)

Run recursive linear search,

Return Error if   results value is equal to negative,   And if it is not, Find the value And display the Value to the User, that they entered and show them the Position of that value in a list.

Task3.
Made A few Adjustments, so its easier to read the Code

"word"  is now "wordStr"   Meaning that it hold one word of string, that is run through the if loop when Splitting getline Sentence or list of words into separate  ones.

Resolved Error:

Functions were listed with  " _ " Which caused a problem and are usually used for Private Protected functions.

Which is not necessary.

Notes :
I thought of making it easier to read, but forgot about the proper use of Underscore.

Error two:
Whole program stopped working with Task3.h  file, where TakeuserString();

Ran into and issue caused by Visual studio.

Fixed by Replacing   ; with ;

Note :

Not entirely sure what it happened, and looking forward to understanding the issue and why or how it can happen.

Commenting >

With the Comments I wanted to make sure that every element is commented so, if one day a programmer comes to see the code, he can easily terminate every part of the code.

The problem can come, as there are a lot of comments that touch or go through Basics, but it Is still as for me, something I want to make sure to have. If I am going to teach someone in the future, it is a good practise to Comment Every Element or a function.

Learning outcomes:
I was Able to understand what Return -1;  and return 1; entirely mean and that they are programmers self defined in returning an error.  And I decided to Include These Elements in part of the Code.