Start of Task5 and Task2.  Task3 and Task1 fixes. Task
Started Task 2:
In task two I'm required to write pseudocode and make addition, multiplying and Subtraction for two matrices.   And then Calculate the given formula..

Currently in the Works, As it has takes time to understand what I am made to do.

Start of Task5
After reading and writing down How bubble sort works, I have implemented a sorting code for the Given array int A[10] = { 2,7,9,4,1,5,3,6,0,8 };

Added A call to Print out The Current Array with tellArray()

Bubble sort
added 2 placeholder elements, count // Which tells us the Loops Pass number.
int elementholder // Which is used in Swap loop to put the element onto hold and then put it to the desired  Array position.

Setup Future, Insertion and Selection sort methods.

Task 1 :
Rewrote The Text, in order for it to not be to long and so it would be more readable.

Wording changes.

Task 3 >
Change Wording when asking User input.

Commented Out Task2
Until it is implemented so I does not distract us in the meanwhile.