Task7, Task8 start.
Start of task 8
I will be taking an extended look into this one, in the next push.

Task 7 Completed.
During the completion of task 7, I found it difficult to carps, how to solve it so I decided that finally I will and try to learn as much as there is for c++ and opperantunities, with watching 10 hours of c++ and making sure that I understand most of the basics.  While I did reading on a few resources, I found various ways of doing the exercise as listed in pseudo code and research behind it.  Finding out that tree sort - Binary search tree. is quite an effective algorithm to sort things, and they could even be used for windows explorer, tab where it goes from c: folder/ folder and etc. Quite an interesting algorithm indeed.

Even though the struggle reading lecture notes, I found myself that all this time It was better for me to instead learn these online while giving me a greater understanding of the problem at hand.

Currently issues that need to be resolved are task 2 and task 5. w
which are still in development

I made some of the function names and other things more clear for myself, this in reality can be good for me reading it but it can also have problems in the future, as the code has too much information like comments, special names that are way longer than they should be.

Removed a few unsessecary thing from task 1 , task 2 as well like Include Array and string as they are already included with the standard.